Fix: utils 폴더 생성 및 hideName, hideEmail 함수 구현
