+++ src/components/Admin/UserCard.tsx
@@ -3,6 +3,8 @@ import styled from 'styled-components'
 import { COLORS, FONTSIZE, FONTWEGHT } from '@src/styles/root'
 import Select from '@components/common/Select'
 import Button from '../common/Button'
+import hideName from '@src/utils/hideName'
+import hideEmail from '@src/utils/hideEmail'
 
 const UserCard = () => {
   const name = '주우민'
@@ -11,43 +13,19 @@ const UserCard = () => {
   const date = '2023-03-24'
   const grade = '일반회원'
 
-  const maskingName = (strName: string) => {
-    if (strName.length > 2) {
-      let originName = strName.split('')
-      originName.forEach(function (name, i) {
-        if (i === 0 || i === originName.length - 1) return
-        originName[i] = '*'
-      })
-      let joinName = originName.join()
-      return joinName.replace(/,/g, '')
-    } else {
-      let pattern = /.$/ // 정규식
-      return strName.replace(pattern, '*')
-    }
-  }
-
-  function hideString(str: string) {
-    if (str.length > 3) {
-      const hiddenChars = str.substring(3).replace(/./g, '*')
-      return str.substring(0, 3) + hiddenChars
-    } else {
-      return str
-    }
-  }
-
   const selectOptions = ['일반회원', '관리자']
 
   return (
     <UserCardStyle>
-      <span>{maskingName(name)}</span>
+      <span>{hideName(name)}</span>
       <span>{nickname}</span>
-      <span>{hideString(email)}</span>
+      <span>{hideEmail(email)}</span>
       <span>{date}</span>
       <span>글/구매평/문의</span>
       <span>
         <Select
           options={selectOptions}
-          initial="일반회원"
+          initial={grade}
           width="108px"
           height="24px"
           borderColor="transparent"
@@ -79,14 +57,5 @@ const UserCardStyle = styled.div`
   span {
     justify-self: center;
     align-self: center;
-    :last-child {
-      /* display: flex;
-      justify-content: center;
-      align-items: center;
-      width: 58px;
-      height: 32px;
-      background-color: ${COLORS.black};
-      color: ${COLORS.white}; */
-    }
   }
 `
